refactor: #28 페이징 쿼리 최적화

+++ src/main/java/org/team1/nbe1_2_team01/domain/board/comment/service/CommentService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface CommentService {
 
-    List<CommentResponse> getReviewsByPage(Long boardId, int page);
+    List<CommentResponse> getReviewsByPage(Long boardId, Long page);
 
     Message deleteById(Long id);
